+++ src/logic/target.js
@@ -209,12 +209,11 @@ Target.forPortfolio = function (portfolio, json) {
   target.portfolio = portfolio
 
   portfolio.target = target
-  target.watch(portfolio, "total", () => target.computeAll())
-
   const XLM = Asset.resolve("XLM")
-  portfolio.trap("minimumBalance", () => {
+  target.watch(portfolio, "minimumBalance", () => {
     XLM.target.min = portfolio.minimumBalance
   })
+  target.watch(portfolio, "total", () => target.computeAll())
 
   // Add/Remove assets after trustline change.
   portfolio.assets.forEach(asset => {